修改range data type 從int 改為double

+++ Yungching.Common/Dto/GetEstate.cs
@@ -7,7 +7,7 @@
         public string Address { get; set; }
         public int Price { get; set; }
         public string Type { get; set; }
-        public int Range { get; set; }
+        public double Range { get; set; }
         public bool Status { get; set; }
     }
